Closes #22: Error when running on non-trivial Oozie XML
Fix a bug in `dot_wrapper.py` file. The bug was that the code there checked
whether the name of a node is equal "INPUT" or "OUTPUT". Such names were
disallowed (an exception was thrown in such case).
However, disallowing it here didn't make sense since these are valid names of
special I/O type nodes. This condition is replaced in this commit with
much more liberal one. The new condition simply makes sure that the
names of the nodes are not duplicated.

The greater problem here was that the code was lacking integration-like
end-to-end tests that would check if the processing starting from Oozie
XML file and ending in generating an image works properly. Because such tests
were absent, the bug was overlooked. In order to fix this greater
problem, the commit introduces such tests as well.